--- recipes/msys2/20161025/conandata.yml
@@ -1,15 +1,0 @@
-sources:
-  "20161025":
-    x86:
-      url: http://repo.msys2.org/distrib/i686/msys2-base-i686-20161025.tar.xz
-      filename: msys2-base-i686-20161025.tar.xz
-    x86_64:
-      url: http://repo.msys2.org/distrib/x86_64/msys2-base-x86_64-20161025.tar.xz
-      filename: msys2-base-x86_64-20161025.tar.xz
-
-checksum:
-  "20161025":
-    x86:
-      sha256: 8bafd3d52f5a51528a8671c1cae5591b36086d6ea5b1e76e17e390965cf6768f
-    x86_64:
-      sha256: bb1f1a0b35b3d96bf9c15092da8ce969a84a134f7b08811292fbc9d84d48c65d
