+++ recipes/msys2/all/LICENSE.md
