+++ recipes/msys2/all/conandata.yml
@@ -1,4 +1,9 @@
 sources:
+  "20161025":
+    - url: http://repo.msys2.org/distrib/i686/msys2-base-i686-20161025.tar.xz
+      sha256: 8bafd3d52f5a51528a8671c1cae5591b36086d6ea5b1e76e17e390965cf6768f
+    - url: http://repo.msys2.org/distrib/x86_64/msys2-base-x86_64-20161025.tar.xz
+      sha256: bb1f1a0b35b3d96bf9c15092da8ce969a84a134f7b08811292fbc9d84d48c65d
   "20190524":
     - url: http://repo.msys2.org/distrib/i686/msys2-base-i686-20190524.tar.xz
       sha256: "7a463afae8bf6ce8262f010a9a1648d056ad5cefc66d6eb69fe948c57d4ccb53"
