+++ recipes/msys2/config.yml
@@ -1,0 +1,5 @@
+versions:
+    "20161025":
+      folder: "all"
+    "20190524":
+      folder: "all"
